Changed id fields from int to long to reflect the entities.

+++ src/main/java/dog/ticketlords/TicketlordsBE/controller/UserController.java
@@ -47,13 +47,13 @@ public class UserController {
   }
 
   @GetMapping("/user/{id}")
-  public ResponseEntity<RegisteredUser> getRegisteredUserById(@PathVariable int id) {
+  public ResponseEntity<RegisteredUser> getRegisteredUserById(@PathVariable long id) {
     Optional<RegisteredUser> user = this.userService.getRegUser(id);
     return user.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
   }
 
   @DeleteMapping("/user/{id}")
-  public ResponseEntity<Void> deleteRegisteredUserById(@PathVariable int id) {
+  public ResponseEntity<Void> deleteRegisteredUserById(@PathVariable long id) {
     if (this.userService.getRegUser(id).isPresent()) {
       this.userService.deleteRegisteredUserFromDatabase(id);
       return ResponseEntity.noContent().build();
